+++ components/Footer.tsx
@@ -1,6 +1,7 @@
+import versionData from "@/constants/version";
 import Link from "next/link";
 
-const Footer = () => {
+const Footer = async () => {
   return (
     <div
       id="footer"
@@ -10,6 +11,9 @@ const Footer = () => {
         <Link href="#about" className="footer-link">
           About
         </Link>
+        <Link href="#experience" className="footer-link">
+          Experience
+        </Link>
         <Link href="#projects" className="footer-link">
           Projects
         </Link>
@@ -18,7 +22,7 @@ const Footer = () => {
         </Link>
       </div>
       <div className="flex flex-col space-y-1 text-center">
-        <small>
+        <p>
           Copyright &copy; {new Date().getFullYear()}{" "}
           <Link
             href="https://github.com/p55d2k"
@@ -28,10 +32,18 @@ const Footer = () => {
             p55d2k
           </Link>{" "}
           (Zhi Kai)
-        </small>
-        <small>
-          This was created in 2024, when I was 14! 🎉
-        </small>
+        </p>
+        <p className="text-sm">This was created in 2024, when I was 14! 🎉</p>
+        <p className="text-sm">
+          <Link
+            href="https://github.com/p55d2k/zk-website"
+            target="_blank"
+            className="text-gray-400 hover:text-gray-300 transition-all ease-in-out duration-300"
+          >
+            Source Code
+          </Link>{" "}
+          • v{versionData.version} • Last updated: {versionData.lastUpdated}
+        </p>
       </div>
     </div>
   );