+++ components/subsections/Story.tsx
@@ -2,7 +2,7 @@
 
 import Image from "next/image";
 
-import SubsectionHeader from "../modules/SubsectionHeader";
+import SubsectionHeader from "../SubsectionHeader";
 
 import { useMedia } from "react-use";